Fix compilation errors, fix session types and auth options export
