+++ src/app/api/cars/sales/route.ts
@@ -2,7 +2,7 @@ import prisma from "@/lib/prisma"
 import { NextResponse } from "next/server"
 import { z } from "zod";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 
 export async function GET() {
     const sales = await prisma.carSale.findMany({})
