+++ src/app/sale/page.tsx
@@ -1,5 +1,4 @@
 import { Button } from "@/components/ui/button";
-import { PlusCircle } from "lucide-react";
 import { JSX } from "react";
 import Link from "next/link";
 import { SaleForm } from "@/components/sale-form";
