+++ src/types/session.ts
@@ -5,5 +5,6 @@ export interface SessionType extends DefaultSession {
     name?: string | null;
     email?: string | null;
     image?: string | null;
+    id?: string | null
   };
 }
